Refactor bootstrap & autoloaders
- Cleaned up main plugin file.
- Created legacy bootstrap for back-compat
- Created new plugin config array.
- Adds new prereq and autoloader packages
- Uses composer autoloading
